requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.1
openai>=1.10.0
python-dotenv>=1.0.0
supabase>=2.0.0
google-api-python-client>=2.111.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
unstructured>=0.10.0
pypdf>=3.17.0
python-docx>=0.8.11
chromadb>=0.4.0
tiktoken>=0.5.0
tabulate>=0.9.0
rich>=13.6.0
psycopg2-binary>=2.9.9 